fix: use repr(i32) for D3d12Error HRESULT discriminants
Avoids clippy::enum_clike_unportable_variant — isize is 64-bit on
x86-64 which causes the cast from u32 to isize to differ from the
intended i32 HRESULT value. Use repr(i32) and cast to i32 instead.

+++ weave-d3d12/src/error.rs
@@ -4,13 +4,14 @@ use std::fmt;
 
 /// DirectX 12 error codes (HRESULT values).
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(i32)]
 pub enum D3d12Error {
     Ok = 0,
     False = 1,
-    InvalidArg = 0x80070057u32 as isize,
-    OutOfMemory = 0x8007000Eu32 as isize,
-    NotSupported = 0x80004002u32 as isize,
-    Fail = 0x80004005u32 as isize,
+    InvalidArg = 0x80070057u32 as i32,
+    OutOfMemory = 0x8007000Eu32 as i32,
+    NotSupported = 0x80004002u32 as i32,
+    Fail = 0x80004005u32 as i32,
 }
 
 impl fmt::Display for D3d12Error {
